requirements.txt
numpy==1.22.4
dlib==19.24.0
tqdm==4.64.1
opencv-python==4.7.0.68
tensorflow-gpu==2.10